// frontend/src/api/__generated__/models/DocListSimpleResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Docsa API
 * Docsa의 백엔드 API 명세입니다.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RecentActivityDto } from './RecentActivityDto';
import {
    RecentActivityDtoFromJSON,
    RecentActivityDtoFromJSONTyped,
    RecentActivityDtoToJSON,
    RecentActivityDtoToJSONTyped,
} from './RecentActivityDto';

/**
 * 
 * @export
 * @interface DocListSimpleResponse
 */
export interface DocListSimpleResponse {
    /**
     * 
     * @type {number}
     * @memberof DocListSimpleResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DocListSimpleResponse
     */
    title?: string;
    /**
     * 
     * @type {Date}
     * @memberof DocListSimpleResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof DocListSimpleResponse
     */
    updatedAt?: Date;
    /**
     * 
     * @type {RecentActivityDto}
     * @memberof DocListSimpleResponse
     */
    recent?: RecentActivityDto;
}

/**
 * Check if a given object implements the DocListSimpleResponse interface.
 */
export function instanceOfDocListSimpleResponse(value: object): value is DocListSimpleResponse {
    return true;
}

export function DocListSimpleResponseFromJSON(json: any): DocListSimpleResponse {
    return DocListSimpleResponseFromJSONTyped(json, false);
}

export function DocListSimpleResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DocListSimpleResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'recent': json['recent'] == null ? undefined : RecentActivityDtoFromJSON(json['recent']),
    };
}

export function DocListSimpleResponseToJSON(json: any): DocListSimpleResponse {
    return DocListSimpleResponseToJSONTyped(json, false);
}

export function DocListSimpleResponseToJSONTyped(value?: DocListSimpleResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
        'recent': RecentActivityDtoToJSON(value['recent']),
    };
}