// frontend/src/api/__generated__/models/PageDocListSimpleResponse.ts
export interface PageDocListSimpleResponse {
    /**
     * 
     * @type {number}
     * @memberof PageDocListSimpleResponse
     */
    totalPages?: number;
    /**
     * 
     * @type {number}
     * @memberof PageDocListSimpleResponse
     */
    totalElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageDocListSimpleResponse
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {number}
     * @memberof PageDocListSimpleResponse
     */
    size?: number;
    /**
     * 
     * @type {Array<DocListSimpleResponse>}
     * @memberof PageDocListSimpleResponse
     */
    content?: Array<DocListSimpleResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageDocListSimpleResponse
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageDocListSimpleResponse
     */
    sort?: SortObject;
    /**
     * 
     * @type {number}
     * @memberof PageDocListSimpleResponse
     */
    numberOfElements?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageDocListSimpleResponse
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageDocListSimpleResponse
     */
    last?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageDocListSimpleResponse
     */
    empty?: boolean;
}